Added edit and delete actions for BanquetMenuItems.
By adding these actions to the controller, we are able to delete a menu item from a banquet.

I have also modified a view to show these menu items in table form.

The viewModel has a comment for traceability. Since this viewModel is very similar to another called "BanquetMenuItemsViewModel" while the one used for this function is "BanquetMenuViewModel". Similar names, different properties within.